Moved interface for LoginNoJetpackFragment to a separate class
