+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/LoginNoJetpackListener.kt
@@ -1,0 +1,13 @@
+package com.woocommerce.android.ui.login
+
+interface LoginNoJetpackListener {
+    fun showJetpackInstructions()
+    fun showWhatIsJetpackDialog()
+    fun showEmailLoginScreen(siteAddress: String?)
+    fun showUsernamePasswordScreen(
+        siteAddress: String?,
+        endpointAddress: String?,
+        inputUsername: String?,
+        inputPassword: String?
+    )
+}
